+++ omopcdmlib/Models/Cohort.cs
@@ -5,9 +5,9 @@ namespace omopcdmlib.Models
 {
     public partial class Cohort
     {
-        public long Id { get; set; }
-        public long CohortDefinitionId { get; set; }
-        public long SubjectId { get; set; }
+        public int Id { get; set; }
+        public int CohortDefinitionId { get; set; }
+        public int SubjectId { get; set; }
         public byte[] CohortStartDate { get; set; }
         public byte[] CohortEndDate { get; set; }
     }
