+++ omopcdmlib/Models/Death.cs
@@ -5,12 +5,12 @@ namespace omopcdmlib.Models
 {
     public partial class Death
     {
-        public long Id { get; set; }
-        public long PersonId { get; set; }
+        public int Id { get; set; }
+        public int PersonId { get; set; }
         public byte[] DeathDate { get; set; }
-        public long DeathTypeConceptId { get; set; }
-        public long? CauseConceptId { get; set; }
+        public int DeathTypeConceptId { get; set; }
+        public int? CauseConceptId { get; set; }
         public string CauseSourceValue { get; set; }
-        public long? CauseSourceConceptId { get; set; }
+        public int? CauseSourceConceptId { get; set; }
     }
 }
