+++ omopcdmlib/Models/DeviceCost.cs
@@ -5,10 +5,10 @@ namespace omopcdmlib.Models
 {
     public partial class DeviceCost
     {
-        public long Id { get; set; }
-        public long DeviceCostId { get; set; }
-        public long DeviceExposureId { get; set; }
-        public long? CurrencyConceptId { get; set; }
+        public int Id { get; set; }
+        public int DeviceCostId { get; set; }
+        public int DeviceExposureId { get; set; }
+        public int? CurrencyConceptId { get; set; }
         public byte[] PaidCopay { get; set; }
         public byte[] PaidCoinsurance { get; set; }
         public byte[] PaidTowardDeductible { get; set; }
@@ -16,6 +16,6 @@ namespace omopcdmlib.Models
         public byte[] PaidByCoordinationBenefits { get; set; }
         public byte[] TotalOutOfPocket { get; set; }
         public byte[] TotalPaid { get; set; }
-        public long? PayerPlanPeriodId { get; set; }
+        public int? PayerPlanPeriodId { get; set; }
     }
 }
