+++ omopcdmlib/Models/PayerPlanPeriod.cs
@@ -5,9 +5,9 @@ namespace omopcdmlib.Models
 {
     public partial class PayerPlanPeriod
     {
-        public long Id { get; set; }
-        public long PayerPlanPeriodId { get; set; }
-        public long PersonId { get; set; }
+        public int Id { get; set; }
+        public int PayerPlanPeriodId { get; set; }
+        public int PersonId { get; set; }
         public byte[] PayerPlanPeriodStartDate { get; set; }
         public byte[] PayerPlanPeriodEndDate { get; set; }
         public string PayerSourceValue { get; set; }
