+++ omopcdmlib/Models/ProcedureOccurrence.cs
@@ -5,18 +5,18 @@ namespace omopcdmlib.Models
 {
     public partial class ProcedureOccurrence
     {
-        public long Id { get; set; }
-        public long ProcedureOccurrenceId { get; set; }
-        public long PersonId { get; set; }
-        public long ProcedureConceptId { get; set; }
+        public int Id { get; set; }
+        public int ProcedureOccurrenceId { get; set; }
+        public int PersonId { get; set; }
+        public int ProcedureConceptId { get; set; }
         public byte[] ProcedureDate { get; set; }
-        public long ProcedureTypeConceptId { get; set; }
-        public long? ModifierConceptId { get; set; }
-        public long? Quantity { get; set; }
-        public long? ProviderId { get; set; }
-        public long? VisitOccurrenceId { get; set; }
+        public int ProcedureTypeConceptId { get; set; }
+        public int? ModifierConceptId { get; set; }
+        public int? Quantity { get; set; }
+        public int? ProviderId { get; set; }
+        public int? VisitOccurrenceId { get; set; }
         public string ProcedureSourceValue { get; set; }
-        public long? ProcedureSourceConceptId { get; set; }
+        public int? ProcedureSourceConceptId { get; set; }
         public string QualifierSourceValue { get; set; }
     }
 }
