+++ omopcdmlib/Models/VisitOccurrence.cs
@@ -5,18 +5,18 @@ namespace omopcdmlib.Models
 {
     public partial class VisitOccurrence
     {
-        public long Id { get; set; }
-        public long VisitOccurrenceId { get; set; }
-        public long PersonId { get; set; }
-        public long VisitConceptId { get; set; }
+        public int Id { get; set; }
+        public int VisitOccurrenceId { get; set; }
+        public int PersonId { get; set; }
+        public int VisitConceptId { get; set; }
         public byte[] VisitStartDate { get; set; }
         public string VisitStartTime { get; set; }
         public byte[] VisitEndDate { get; set; }
         public string VisitEndTime { get; set; }
-        public long VisitTypeConceptId { get; set; }
-        public long? ProviderId { get; set; }
-        public long? CareSiteId { get; set; }
+        public int VisitTypeConceptId { get; set; }
+        public int? ProviderId { get; set; }
+        public int? CareSiteId { get; set; }
         public string VisitSourceValue { get; set; }
-        public long? VisitSourceConceptId { get; set; }
+        public int? VisitSourceConceptId { get; set; }
     }
 }
